requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djangorestframework-simplejwt[blacklist]
psycopg2
requests
python-json-logger